@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+5
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -26,116 +26,132 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   auto_size_text: ^3.0.0
-  built_collection: 5.1.1
-  built_value: 8.4.3
   cached_network_image: 3.2.1
   calendar_date_picker2: ^0.3.7
-  cloud_firestore: 4.2.0
-  cloud_firestore_platform_interface: 5.9.1
-  cloud_firestore_web: 3.1.1
-  cloud_functions: 4.0.6
-  cloud_functions_platform_interface: 5.1.25
-  cloud_functions_web: 4.3.14
-  collection: 1.17.0
+  cloud_firestore: 4.8.0
+  cloud_firestore_platform_interface: 5.15.0
+  cloud_firestore_web: 3.6.0
+  cloud_functions: 4.3.2
+  cloud_functions_platform_interface: 5.4.2
+  cloud_functions_web: 4.5.2
+  collection: 1.17.2
   dropdown_button2: 2.0.0
   easy_debounce: 2.0.1
   excel: ^2.0.1
   expandable: 5.0.1
-  file_picker: 4.5.1
+  file_picker: 5.2.6
   firebase_auth: ^4.2.9
   firebase_auth_platform_interface: ^6.11.11
   firebase_auth_web: ^5.2.8
   firebase_core: ^2.7.0
   firebase_core_platform_interface: ^4.5.3
   firebase_core_web: ^2.2.1
-  firebase_messaging: 14.1.4
-  firebase_messaging_platform_interface: 4.2.8
-  firebase_messaging_web: 3.2.9
-  firebase_storage: 11.0.7
-  firebase_storage_platform_interface: 4.1.25
-  firebase_storage_web: 3.3.17
+  firebase_messaging: 14.6.2
+  firebase_messaging_platform_interface: 4.5.2
+  firebase_messaging_web: 3.5.2
+  firebase_storage: 11.2.2
+  firebase_storage_platform_interface: 4.4.2
+  firebase_storage_web: 3.6.2
   fl_chart: 0.55.0
-  flutter_animate: 1.0.0
+  flutter_animate: 4.1.1+1
   flutter_cache_manager: 3.3.0
-  flutter_google_places: 0.3.0
+  flutter_google_places: 
+    git:
+      url: https://github.com/fluttercommunity/flutter_google_places
+      ref: v0.3.2
   flutter_html: 3.0.0-alpha.5
+  flutter_plugin_android_lifecycle: 2.0.9
   flutter_rating_bar: 4.0.1
-  flutter_svg: 2.0.3
+  flutter_svg: 2.0.4
   font_awesome_flutter: ^10.4.0
   from_css_color: 2.0.0
-  geolocator: 9.0.1
-  geolocator_android: 4.1.0
-  geolocator_apple: 2.2.1
-  geolocator_platform_interface: 4.0.6
+  geolocator: 10.0.1
+  geolocator_android: 4.2.4
+  geolocator_apple: 2.2.7
+  geolocator_platform_interface: 4.0.8
   geolocator_web: 2.1.6
-  go_router: 3.1.0
-  google_fonts: 3.0.1
-  google_maps: 6.2.0
-  google_maps_flutter: 2.1.8
-  google_maps_flutter_platform_interface: 2.2.1
-  google_maps_flutter_web: 0.4.0+1
-  google_sign_in: 5.4.0
-  google_sign_in_ios: 5.4.0
-  google_sign_in_platform_interface: 2.2.0
-  google_sign_in_web: 0.10.2
-  image_picker: 0.8.5+3
-  image_picker_android: 0.8.5+3
-  image_picker_for_web: 2.1.10
-  image_picker_ios: 0.8.6+1
-  image_picker_platform_interface: 2.6.2
-  infinite_scroll_pagination: 3.2.0
-  intl: 0.17.0
-  json_path: 0.4.1
-  json_serializable: 6.6.1
+  go_router: 7.1.1
+  google_fonts: 4.0.3
+  google_maps: 6.3.0
+  google_maps_flutter: 2.2.6
+  google_maps_flutter_platform_interface: 2.2.6
+  google_maps_flutter_web: 0.4.0+8
+  google_sign_in: 6.0.2
+  google_sign_in_android: 6.1.8
+  google_sign_in_ios: 5.6.1
+  google_sign_in_platform_interface: 2.4.0
+  google_sign_in_web: 0.11.0+2
+  image_picker: 1.0.4
+  image_picker_android: 0.8.8
+  image_picker_for_web: 3.0.1
+  image_picker_ios: 0.8.8+2
+  image_picker_platform_interface: 2.9.1
+  infinite_scroll_pagination: 
+    git:
+      url: https://github.com/FlutterFlow/infinite_scroll_pagination.git
+      ref: develop
+  intl: 0.18.1
+  json_path: 0.6.2
   mask_text_input_formatter: 2.4.0
   mime_type: 1.0.0
+  package_info_plus: 3.1.0
+  package_info_plus_platform_interface: 2.0.1
   page_transition: 2.0.4
+  path_provider: 2.0.14
+  path_provider_android: 2.0.25
+  path_provider_foundation: 2.2.2
+  path_provider_platform_interface: 2.0.6
   percent_indicator: 4.2.2
   pin_code_fields: 7.4.0
   plugin_platform_interface: 2.1.3
+  pointer_interceptor: 0.9.3+4
   provider: 6.0.4
-  rxdart: 0.27.4
+  rxdart: 0.27.7
   shared_preferences: 2.0.15
-  sign_in_with_apple: 4.1.0
+  shared_preferences_android: 2.1.0
+  shared_preferences_ios: 2.1.1
+  shared_preferences_platform_interface: 2.2.0
+  shared_preferences_web: 2.1.0
+  sign_in_with_apple: 4.3.0
   sign_in_with_apple_platform_interface: 1.0.0
   sign_in_with_apple_web: 1.0.1
-  smooth_page_indicator: 1.0.0+2
-  sqflite: 2.2.0+2
+  smooth_page_indicator: 1.0.1
+  sqflite: 2.2.6
+  stream_transform: 2.1.0
   substring_highlight: 1.0.33
-  table_calendar: 3.0.6
+  table_calendar: 3.0.9
   timeago: 3.2.2
-  url_launcher: 6.1.5
-  video_player: 2.4.5
-  video_player_android: 2.3.8
-  video_player_avfoundation: 2.3.5
-  video_player_platform_interface: 5.1.3
-  video_player_web: 2.0.12
+  url_launcher: 6.1.10
+  url_launcher_android: 6.0.27
+  url_launcher_ios: 6.1.4
+  url_launcher_platform_interface: 2.1.2
+  video_player: 2.6.0
+  video_player_android: 2.4.3
+  video_player_avfoundation: 2.4.2
+  video_player_platform_interface: 6.1.0
+  video_player_web: 2.0.15
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
 
-dependency_overrides:
-  provider: 6.0.0
-  rxdart: 0.27.4
-  webview_flutter_android: 2.10.3
-  webview_flutter_platform_interface: 1.9.5
-  webview_flutter_wkwebview: 2.9.5
-
 dev_dependencies:
-  build_runner: 2.3.3
-  built_value_generator: 8.4.3
-  flutter_launcher_icons: 0.12.0
+  flutter_launcher_icons: 0.13.1
+  flutter_lints: 3.0.0
+  image: 4.0.17
+  lints: 3.0.0
 
   flutter_test:
     sdk: flutter
 
 
-flutter_icons:
+flutter_launcher_icons:
   android: true
   ios: true
   remove_alpha_ios: true
+  web:
+    generate: true
   image_path: 'assets/images/app_launcher_icon.png'
 
 
