@@ -1,44 +1,67 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart';
+
+import '/backend/schema/util/firestore_util.dart';
+import '/backend/schema/util/schema_util.dart';
+
 import 'index.dart';
-import 'serializers.dart';
-import 'package:built_value/built_value.dart';
+import '/flutter_flow/flutter_flow_util.dart';
 
-part 'promotion_record.g.dart';
+class PromotionRecord extends FirestoreRecord {
+  PromotionRecord._(
+    DocumentReference reference,
+    Map<String, dynamic> data,
+  ) : super(reference, data) {
+    _initializeFields();
+  }
 
-abstract class PromotionRecord
-    implements Built<PromotionRecord, PromotionRecordBuilder> {
-  static Serializer<PromotionRecord> get serializer =>
-      _$promotionRecordSerializer;
+  // "title" field.
+  String? _title;
+  String get title => _title ?? '';
+  bool hasTitle() => _title != null;
 
-  String? get title;
+  // "subtitle" field.
+  String? _subtitle;
+  String get subtitle => _subtitle ?? '';
+  bool hasSubtitle() => _subtitle != null;
 
-  String? get subtitle;
+  // "img" field.
+  String? _img;
+  String get img => _img ?? '';
+  bool hasImg() => _img != null;
 
-  String? get img;
+  // "city_link" field.
+  DocumentReference? _cityLink;
+  DocumentReference? get cityLink => _cityLink;
+  bool hasCityLink() => _cityLink != null;
 
-  @BuiltValueField(wireName: 'city_link')
-  DocumentReference? get cityLink;
+  // "status" field.
+  String? _status;
+  String get status => _status ?? '';
+  bool hasStatus() => _status != null;
 
-  String? get status;
+  // "moderation" field.
+  bool? _moderation;
+  bool get moderation => _moderation ?? false;
+  bool hasModeration() => _moderation != null;
 
-  bool? get moderation;
-
-  bool? get top;
-
-  @BuiltValueField(wireName: kDocumentReferenceField)
-  DocumentReference? get ffRef;
-  DocumentReference get reference => ffRef!;
+  // "top" field.
+  bool? _top;
+  bool get top => _top ?? false;
+  bool hasTop() => _top != null;
 
   DocumentReference get parentReference => reference.parent.parent!;
 
-  static void _initializeBuilder(PromotionRecordBuilder builder) => builder
-    ..title = ''
-    ..subtitle = ''
-    ..img = ''
-    ..status = ''
-    ..moderation = false
-    ..top = false;
+  void _initializeFields() {
+    _title = snapshotData['title'] as String?;
+    _subtitle = snapshotData['subtitle'] as String?;
+    _img = snapshotData['img'] as String?;
+    _cityLink = snapshotData['city_link'] as DocumentReference?;
+    _status = snapshotData['status'] as String?;
+    _moderation = snapshotData['moderation'] as bool?;
+    _top = snapshotData['top'] as bool?;
+  }
 
   static Query<Map<String, dynamic>> collection([DocumentReference? parent]) =>
       parent != null
@@ -48,22 +71,35 @@ abstract class PromotionRecord
   static DocumentReference createDoc(DocumentReference parent) =>
       parent.collection('promotion').doc();
 
-  static Stream<PromotionRecord> getDocument(DocumentReference ref) => ref
-      .snapshots()
-      .map((s) => serializers.deserializeWith(serializer, serializedData(s))!);
+  static Stream<PromotionRecord> getDocument(DocumentReference ref) =>
+      ref.snapshots().map((s) => PromotionRecord.fromSnapshot(s));
 
-  static Future<PromotionRecord> getDocumentOnce(DocumentReference ref) => ref
-      .get()
-      .then((s) => serializers.deserializeWith(serializer, serializedData(s))!);
+  static Future<PromotionRecord> getDocumentOnce(DocumentReference ref) =>
+      ref.get().then((s) => PromotionRecord.fromSnapshot(s));
 
-  PromotionRecord._();
-  factory PromotionRecord([void Function(PromotionRecordBuilder) updates]) =
-      _$PromotionRecord;
+  static PromotionRecord fromSnapshot(DocumentSnapshot snapshot) =>
+      PromotionRecord._(
+        snapshot.reference,
+        mapFromFirestore(snapshot.data() as Map<String, dynamic>),
+      );
 
   static PromotionRecord getDocumentFromData(
-          Map<String, dynamic> data, DocumentReference reference) =>
-      serializers.deserializeWith(serializer,
-          {...mapFromFirestore(data), kDocumentReferenceField: reference})!;
+    Map<String, dynamic> data,
+    DocumentReference reference,
+  ) =>
+      PromotionRecord._(reference, mapFromFirestore(data));
+
+  @override
+  String toString() =>
+      'PromotionRecord(reference: ${reference.path}, data: $snapshotData)';
+
+  @override
+  int get hashCode => reference.path.hashCode;
+
+  @override
+  bool operator ==(other) =>
+      other is PromotionRecord &&
+      reference.path.hashCode == other.reference.path.hashCode;
 }
 
 Map<String, dynamic> createPromotionRecordData({
@@ -75,19 +111,46 @@ Map<String, dynamic> createPromotionRecordData({
   bool? moderation,
   bool? top,
 }) {
-  final firestoreData = serializers.toFirestore(
-    PromotionRecord.serializer,
-    PromotionRecord(
-      (p) => p
-        ..title = title
-        ..subtitle = subtitle
-        ..img = img
-        ..cityLink = cityLink
-        ..status = status
-        ..moderation = moderation
-        ..top = top,
-    ),
+  final firestoreData = mapToFirestore(
+    <String, dynamic>{
+      'title': title,
+      'subtitle': subtitle,
+      'img': img,
+      'city_link': cityLink,
+      'status': status,
+      'moderation': moderation,
+      'top': top,
+    }.withoutNulls,
   );
 
   return firestoreData;
+}
+
+class PromotionRecordDocumentEquality implements Equality<PromotionRecord> {
+  const PromotionRecordDocumentEquality();
+
+  @override
+  bool equals(PromotionRecord? e1, PromotionRecord? e2) {
+    return e1?.title == e2?.title &&
+        e1?.subtitle == e2?.subtitle &&
+        e1?.img == e2?.img &&
+        e1?.cityLink == e2?.cityLink &&
+        e1?.status == e2?.status &&
+        e1?.moderation == e2?.moderation &&
+        e1?.top == e2?.top;
+  }
+
+  @override
+  int hash(PromotionRecord? e) => const ListEquality().hash([
+        e?.title,
+        e?.subtitle,
+        e?.img,
+        e?.cityLink,
+        e?.status,
+        e?.moderation,
+        e?.top
+      ]);
+
+  @override
+  bool isValidKey(Object? o) => o is PromotionRecord;
 }
