@@ -1,20 +1,23 @@
-import '/auth/auth_util.dart';
+import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
 import '/flutter_flow/flutter_flow_drop_down.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/form_field_controller.dart';
+import 'admin_edit_stuff_status_widget.dart' show AdminEditStuffStatusWidget;
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 
-class AdminEditStuffStatusModel extends FlutterFlowModel {
+class AdminEditStuffStatusModel
+    extends FlutterFlowModel<AdminEditStuffStatusWidget> {
   ///  State fields for stateful widgets in this component.
 
   // State field(s) for DropDown667 widget.
   String? dropDown667Value;
-  FormFieldController<String>? dropDown667Controller;
+  FormFieldController<String>? dropDown667ValueController;
 
   /// Initialization and disposal methods.
 
@@ -22,6 +25,7 @@ class AdminEditStuffStatusModel extends FlutterFlowModel {
 
   void dispose() {}
 
-  /// Additional helper methods are added here.
+  /// Action blocks are added here.
 
+  /// Additional helper methods are added here.
 }
