@@ -1,4 +1,4 @@
-import '/auth/auth_util.dart';
+import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
 import '/flutter_flow/flutter_flow_drop_down.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
@@ -6,6 +6,7 @@ import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/form_field_controller.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import 'admin_edit_stuff_status_model.dart';
@@ -54,25 +55,25 @@ class _AdminEditStuffStatusWidgetState
     context.watch<FFAppState>();
 
     return FlutterFlowDropDown<String>(
-      controller: _model.dropDown667Controller ??= FormFieldController<String>(
-        _model.dropDown667Value ??= widget.user!.adminStatus,
+      controller: _model.dropDown667ValueController ??=
+          FormFieldController<String>(
+        _model.dropDown667Value ??= widget.user?.adminStatus,
       ),
       options: ['Активен', 'Не активен'],
       onChanged: (val) async {
         setState(() => _model.dropDown667Value = val);
-        final userUpdateData = createUserRecordData(
+        await widget.user!.reference.update(createUserRecordData(
           adminStatus: _model.dropDown667Value,
-        );
-        await widget.user!.reference.update(userUpdateData);
+        ));
       },
       width: 180.0,
       height: 50.0,
       textStyle: FlutterFlowTheme.of(context).bodyMedium.override(
             fontFamily: 'Roboto',
             color: () {
-              if (widget.user!.adminStatus == 'Не активен') {
+              if (widget.user?.adminStatus == 'Не активен') {
                 return Color(0xFFE20000);
-              } else if (widget.user!.adminStatus == 'Активен') {
+              } else if (widget.user?.adminStatus == 'Активен') {
                 return FlutterFlowTheme.of(context).green;
               } else {
                 return FlutterFlowTheme.of(context).primaryText;
@@ -94,6 +95,7 @@ class _AdminEditStuffStatusWidgetState
       margin: EdgeInsetsDirectional.fromSTEB(12.0, 4.0, 12.0, 4.0),
       hidesUnderline: true,
       isSearchable: false,
+      isMultiSelect: false,
     );
   }
 }
