@@ -1,16 +1,19 @@
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
+import 'empty_services_widget.dart' show EmptyServicesWidget;
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 
-class EmptyServicesModel extends FlutterFlowModel {
+class EmptyServicesModel extends FlutterFlowModel<EmptyServicesWidget> {
   /// Initialization and disposal methods.
 
   void initState(BuildContext context) {}
 
   void dispose() {}
 
-  /// Additional helper methods are added here.
+  /// Action blocks are added here.
 
+  /// Additional helper methods are added here.
 }
