@@ -7,8 +7,6 @@ import 'package:intl/intl.dart';
 export 'package:fl_chart/fl_chart.dart'
     show BarAreaData, FlDotData, LineChartBarData, BarChartAlignment;
 
-final _format = (NumberFormat()..significantDigits = 2);
-
 class FlutterFlowLineChart extends StatelessWidget {
   const FlutterFlowLineChart({
     Key? key,
@@ -48,10 +46,6 @@ class FlutterFlowLineChart extends StatelessWidget {
           titlesData: getTitlesData(
             xAxisLabelInfo,
             yAxisLabelInfo,
-            (val, _) => Text(
-              _format.format(val),
-              style: xAxisLabelInfo.labelTextStyle,
-            ),
           ),
           lineBarsData: dataWithSpots,
           minX: axisBounds.minX,
@@ -174,7 +168,7 @@ class FlutterFlowBarChart extends StatelessWidget {
         titlesData: getTitlesData(
           xAxisLabelInfo,
           yAxisLabelInfo,
-          (val, _) => Text(
+          getXTitlesWidget: (val, _) => Text(
             xLabels[val.toInt()],
             style: xAxisLabelInfo.labelTextStyle,
           ),
@@ -203,6 +197,7 @@ class FlutterFlowPieChart extends StatelessWidget {
     this.donutHoleColor = Colors.transparent,
     this.sectionLabelType = PieChartSectionLabelType.none,
     this.sectionLabelStyle,
+    this.labelFormatter = const LabelFormatter(),
   }) : super(key: key);
 
   final FFPieChartData data;
@@ -210,6 +205,7 @@ class FlutterFlowPieChart extends StatelessWidget {
   final Color donutHoleColor;
   final PieChartSectionLabelType sectionLabelType;
   final TextStyle? sectionLabelStyle;
+  final LabelFormatter labelFormatter;
 
   double get sumOfValues => data.data.reduce((a, b) => a + b);
 
@@ -228,10 +224,11 @@ class FlutterFlowPieChart extends StatelessWidget {
               final otherPropsLength = data.radius.length;
               switch (sectionLabelType) {
                 case PieChartSectionLabelType.value:
-                  title = _format.format(sectionData);
+                  title = formatLabel(labelFormatter, sectionData);
                   break;
                 case PieChartSectionLabelType.percent:
-                  title = '${_format.format(sectionData / sumOfValues * 100)}%';
+                  title =
+                      '${formatLabel(labelFormatter, sectionData / sumOfValues * 100)}%';
                   break;
                 default:
                   break;
@@ -367,6 +364,7 @@ class AxisLabelInfo {
     this.showLabels = false,
     this.labelTextStyle,
     this.labelInterval,
+    this.labelFormatter = const LabelFormatter(),
   });
 
   final String title;
@@ -374,6 +372,16 @@ class AxisLabelInfo {
   final bool showLabels;
   final TextStyle? labelTextStyle;
   final double? labelInterval;
+  final LabelFormatter labelFormatter;
+}
+
+class LabelFormatter {
+  const LabelFormatter({
+    this.numberFormat,
+  });
+
+  final String Function(double)? numberFormat;
+  NumberFormat get defaultFormat => NumberFormat()..significantDigits = 2;
 }
 
 class AxisBounds {
@@ -477,9 +485,9 @@ List<double?> _dataToDouble(List<dynamic> data) {
 
 FlTitlesData getTitlesData(
   AxisLabelInfo xAxisLabelInfo,
-  AxisLabelInfo yAxisLabelInfo,
+  AxisLabelInfo yAxisLabelInfo, {
   Widget Function(double, TitleMeta)? getXTitlesWidget,
-) =>
+}) =>
     FlTitlesData(
       bottomTitles: AxisTitles(
         axisNameWidget: xAxisLabelInfo.title.isEmpty
@@ -492,8 +500,13 @@ FlTitlesData getTitlesData(
             ? xAxisLabelInfo.titleTextStyle!.fontSize! + 12
             : null,
         sideTitles: SideTitles(
+          getTitlesWidget: (val, _) => getXTitlesWidget != null
+              ? getXTitlesWidget(val, _)
+              : Text(
+                  formatLabel(xAxisLabelInfo.labelFormatter, val),
+                  style: xAxisLabelInfo.labelTextStyle,
+                ),
           showTitles: xAxisLabelInfo.showLabels,
-          getTitlesWidget: getXTitlesWidget,
           interval: xAxisLabelInfo.labelInterval,
         ),
       ),
@@ -511,7 +524,7 @@ FlTitlesData getTitlesData(
             : null,
         sideTitles: SideTitles(
           getTitlesWidget: (val, _) => Text(
-            _format.format(val),
+            formatLabel(yAxisLabelInfo.labelFormatter, val),
             style: yAxisLabelInfo.labelTextStyle,
           ),
           showTitles: yAxisLabelInfo.showLabels,
@@ -519,3 +532,10 @@ FlTitlesData getTitlesData(
         ),
       ),
     );
+
+String formatLabel(LabelFormatter formatter, double value) {
+  if (formatter.numberFormat != null) {
+    return formatter.numberFormat!(value);
+  }
+  return formatter.defaultFormat.format(value);
+}
