@@ -2,21 +2,23 @@ import 'package:dropdown_button2/dropdown_button2.dart';
 
 import 'form_field_controller.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 
 class FlutterFlowDropDown<T> extends StatefulWidget {
   const FlutterFlowDropDown({
+    Key? key,
     required this.controller,
     this.hintText,
     this.searchHintText,
     required this.options,
     this.optionLabels,
-    required this.onChanged,
+    this.onChanged,
+    this.onChangedForMultiSelect,
     this.icon,
     this.width,
     this.height,
     this.fillColor,
     this.searchHintTextStyle,
+    this.searchCursorColor,
     required this.textStyle,
     required this.elevation,
     required this.borderWidth,
@@ -26,19 +28,22 @@ class FlutterFlowDropDown<T> extends StatefulWidget {
     this.hidesUnderline = false,
     this.disabled = false,
     this.isSearchable = false,
-  });
+    this.isMultiSelect = false,
+  }) : super(key: key);
 
   final FormFieldController<T> controller;
   final String? hintText;
   final String? searchHintText;
   final List<T> options;
   final List<String>? optionLabels;
-  final Function(T?) onChanged;
+  final Function(T?)? onChanged;
+  final Function(List<T>?)? onChangedForMultiSelect;
   final Widget? icon;
   final double? width;
   final double? height;
   final Color? fillColor;
   final TextStyle? searchHintTextStyle;
+  final Color? searchCursorColor;
   final TextStyle textStyle;
   final double elevation;
   final double borderWidth;
@@ -48,6 +53,7 @@ class FlutterFlowDropDown<T> extends StatefulWidget {
   final bool hidesUnderline;
   final bool disabled;
   final bool isSearchable;
+  final bool isMultiSelect;
 
   @override
   State<FlutterFlowDropDown<T>> createState() => _FlutterFlowDropDownState<T>();
@@ -56,8 +62,9 @@ class FlutterFlowDropDown<T> extends StatefulWidget {
 class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
   final TextEditingController _textEditingController = TextEditingController();
 
-  void Function() get listener =>
-      () => widget.onChanged(widget.controller.value);
+  void Function() get listener => widget.isMultiSelect
+      ? () {}
+      : () => widget.onChanged!(widget.controller.value);
 
   @override
   void initState() {
@@ -71,22 +78,27 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
     super.dispose();
   }
 
+  List<T> selectedItems = [];
+
   @override
   Widget build(BuildContext context) {
+    final optionToDisplayValue = Map.fromEntries(
+      widget.options.asMap().entries.map((option) => MapEntry(
+          option.value,
+          widget.optionLabels == null ||
+                  widget.optionLabels!.length < option.key + 1
+              ? option.value.toString()
+              : widget.optionLabels![option.key])),
+    );
     final value = widget.options.contains(widget.controller.value)
         ? widget.controller.value
         : null;
     final items = widget.options
-        .asMap()
-        .entries
         .map(
           (option) => DropdownMenuItem<T>(
-            value: option.value,
+            value: option,
             child: Text(
-              widget.optionLabels == null ||
-                      widget.optionLabels!.length < option.key + 1
-                  ? option.value.toString()
-                  : widget.optionLabels![option.key],
+              optionToDisplayValue[option] ?? '',
               style: widget.textStyle,
             ),
           ),
@@ -95,11 +107,26 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
     final hintText = widget.hintText != null
         ? Text(widget.hintText!, style: widget.textStyle)
         : null;
-    void Function(T?)? onChanged =
-        !widget.disabled ? (value) => widget.controller.value = value : null;
-    final dropdownWidget = widget.isSearchable
-        ? _buildSearchableDropdown(value, items, onChanged, hintText)
-        : _buildNonSearchableDropdown(value, items, onChanged, hintText);
+    void Function(T?)? onChanged = widget.disabled || widget.isMultiSelect
+        ? null
+        : (value) => widget.controller.value = value;
+    final dropdownWidget = widget.isMultiSelect
+        ? _buildMultiSelectDropdown(
+            hintText,
+            optionToDisplayValue,
+            widget.isSearchable,
+            widget.onChangedForMultiSelect!,
+            widget.disabled,
+          )
+        : widget.isSearchable
+            ? _buildSearchableDropdown(
+                value,
+                items,
+                onChanged,
+                hintText,
+                optionToDisplayValue,
+              )
+            : _buildNonSearchableDropdown(value, items, onChanged, hintText);
     final childWidget = DecoratedBox(
       decoration: BoxDecoration(
         borderRadius: BorderRadius.circular(widget.borderRadius),
@@ -150,6 +177,7 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
     List<DropdownMenuItem<T>>? items,
     void Function(T?)? onChanged,
     Text? hintText,
+    Map<T, String> optionLabels,
   ) {
     final overlayColor = MaterialStateProperty.resolveWith<Color?>((states) =>
         states.contains(MaterialState.focused) ? Colors.transparent : null);
@@ -190,6 +218,7 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
             expands: true,
             maxLines: null,
             controller: _textEditingController,
+            cursorColor: widget.searchCursorColor,
             decoration: InputDecoration(
               isDense: true,
               contentPadding: const EdgeInsets.symmetric(
@@ -205,19 +234,158 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
           ),
         ),
         searchMatchFn: (item, searchValue) {
-          return item.value
-              .toString()
+          return (optionLabels[item.value] ?? '')
               .toLowerCase()
               .contains(searchValue.toLowerCase());
         },
       ),
 
-      //This to clear the search value when you close the menu
+      // This to clear the search value when you close the menu
       onMenuStateChange: (isOpen) {
         if (!isOpen) {
           _textEditingController.clear();
         }
       },
+    );
+  }
+
+  Widget _buildMultiSelectDropdown(
+    Text? hintText,
+    Map<T, String> optionLabels,
+    bool isSearchable,
+    Function(List<T>?) onChangedForMultiSelect,
+    bool disabled,
+  ) {
+    final overlayColor = MaterialStateProperty.resolveWith<Color?>((states) =>
+        states.contains(MaterialState.focused) ? Colors.transparent : null);
+    final iconStyleData = widget.icon != null
+        ? IconStyleData(icon: widget.icon!)
+        : const IconStyleData();
+    return DropdownButton2<T>(
+      value: selectedItems.isEmpty ? null : selectedItems.last,
+      hint: hintText,
+      items: widget.options.map((item) {
+        return DropdownMenuItem(
+          value: item,
+          // Disable default onTap to avoid closing menu when selecting an item
+          enabled: false,
+          child: StatefulBuilder(
+            builder: (context, menuSetState) {
+              final isSelected = selectedItems.contains(item);
+              return InkWell(
+                onTap: () {
+                  isSelected
+                      ? selectedItems.remove(item)
+                      : selectedItems.add(item);
+                  onChangedForMultiSelect(selectedItems);
+                  //This rebuilds the StatefulWidget to update the button's text
+                  setState(() {});
+                  //This rebuilds the dropdownMenu Widget to update the check mark
+                  menuSetState(() {});
+                },
+                child: Container(
+                  height: double.infinity,
+                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                  child: Row(
+                    children: [
+                      if (isSelected)
+                        const Icon(Icons.check_box_outlined)
+                      else
+                        const Icon(Icons.check_box_outline_blank),
+                      const SizedBox(width: 16),
+                      Expanded(
+                        child: Text(
+                          item as String,
+                          style: widget.textStyle,
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              );
+            },
+          ),
+        );
+      }).toList(),
+      iconStyleData: iconStyleData,
+      buttonStyleData: ButtonStyleData(
+        elevation: widget.elevation!.toInt(),
+        overlayColor: overlayColor,
+      ),
+      menuItemStyleData: MenuItemStyleData(overlayColor: overlayColor),
+      dropdownStyleData: DropdownStyleData(
+        elevation: widget.elevation!.toInt(),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(4),
+          color: widget.fillColor,
+        ),
+      ),
+      // onChanged is handled by the onChangedForMultiSelect function
+      onChanged: disabled ? null : (val) {},
+      isExpanded: true,
+      selectedItemBuilder: (context) {
+        return widget.options.map(
+          (item) {
+            return Container(
+              alignment: AlignmentDirectional.center,
+              child: Text(
+                selectedItems.join(', '),
+                style: const TextStyle(
+                  fontSize: 14,
+                  overflow: TextOverflow.ellipsis,
+                ),
+                maxLines: 1,
+              ),
+            );
+          },
+        ).toList();
+      },
+      dropdownSearchData: isSearchable
+          ? DropdownSearchData<T>(
+              searchController: _textEditingController,
+              searchInnerWidgetHeight: 50,
+              searchInnerWidget: Container(
+                height: 50,
+                padding: const EdgeInsets.only(
+                  top: 8,
+                  bottom: 4,
+                  right: 8,
+                  left: 8,
+                ),
+                child: TextFormField(
+                  expands: true,
+                  maxLines: null,
+                  controller: _textEditingController,
+                  cursorColor: widget.searchCursorColor,
+                  decoration: InputDecoration(
+                    isDense: true,
+                    contentPadding: const EdgeInsets.symmetric(
+                      horizontal: 10,
+                      vertical: 8,
+                    ),
+                    hintText: widget.searchHintText,
+                    hintStyle: widget.searchHintTextStyle,
+                    border: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(8),
+                    ),
+                  ),
+                ),
+              ),
+              searchMatchFn: (item, searchValue) {
+                return (optionLabels[item.value] ?? '')
+                    .toLowerCase()
+                    .contains(searchValue.toLowerCase());
+              },
+            )
+          : null,
+      // This to clear the search value when you close the menu
+      onMenuStateChange: isSearchable
+          ? (isOpen) {
+              if (!isOpen) {
+                _textEditingController.clear();
+              }
+            }
+          : null,
     );
   }
 }
