@@ -34,7 +34,7 @@ T createModel<T extends FlutterFlowModel>(
   return model;
 }
 
-abstract class FlutterFlowModel {
+abstract class FlutterFlowModel<W extends Widget> {
   // Initialization methods
   bool _isInitialized = false;
   void initState(BuildContext context);
@@ -43,7 +43,15 @@ abstract class FlutterFlowModel {
       initState(context);
       _isInitialized = true;
     }
+    if (context.widget is W) _widget = context.widget as W;
   }
+
+  // The widget associated with this model. This is useful for accessing the
+  // parameters of the widget, for example.
+  W? _widget;
+  // This will always be non-null when used, but is nullable to allow us to
+  // dispose of the widget in the [dispose] method (for garbage collection).
+  W get widget => _widget!;
 
   // Dispose methods
   // Whether to dispose this model when the corresponding widget is
@@ -55,6 +63,8 @@ abstract class FlutterFlowModel {
     if (disposeOnWidgetDisposal) {
       dispose();
     }
+    // Remove reference to widget for garbage collection purposes.
+    _widget = null;
   }
 
   // Whether to update the containing page / component on updates.
@@ -102,13 +112,13 @@ class FlutterFlowDynamicModels<T extends FlutterFlowModel> {
         .toList();
   }
 
-  S? getValueAtIndex<S>(int index, S Function(T) getValue) {
+  S? getValueAtIndex<S>(int index, S? Function(T) getValue) {
     final uniqueKey =
         _childrenIndexes.entries.firstWhereOrNull((e) => e.value == index)?.key;
     return getValueForKey(uniqueKey, getValue);
   }
 
-  S? getValueForKey<S>(String? uniqueKey, S Function(T) getValue) {
+  S? getValueForKey<S>(String? uniqueKey, S? Function(T) getValue) {
     final model = _childrenModels[uniqueKey];
     return model != null ? getValue(model) : null;
   }
