@@ -7,19 +7,27 @@ class FFUploadedFile {
     this.bytes,
     this.height,
     this.width,
+    this.blurHash,
   });
 
   final String? name;
   final Uint8List? bytes;
   final double? height;
   final double? width;
+  final String? blurHash;
 
   @override
   String toString() =>
-      'FFUploadedFile(name: $name, bytes: ${bytes?.length ?? 0}, height: $height, width: $width)';
+      'FFUploadedFile(name: $name, bytes: ${bytes?.length ?? 0}, height: $height, width: $width, blurHash: $blurHash,)';
 
   String serialize() => jsonEncode(
-        {'name': name, 'bytes': bytes, 'height': height, 'width': width},
+        {
+          'name': name,
+          'bytes': bytes,
+          'height': height,
+          'width': width,
+          'blurHash': blurHash,
+        },
       );
 
   static FFUploadedFile deserialize(String val) {
@@ -29,17 +37,25 @@ class FFUploadedFile {
       'bytes': serializedData['bytes'] ?? Uint8List.fromList([]),
       'height': serializedData['height'],
       'width': serializedData['width'],
+      'blurHash': serializedData['blurHash'],
     };
     return FFUploadedFile(
       name: data['name'] as String,
-      bytes: data['bytes'] as Uint8List,
+      bytes: Uint8List.fromList(data['bytes'].cast<int>().toList()),
       height: data['height'] as double?,
       width: data['width'] as double?,
+      blurHash: data['blurHash'] as String?,
     );
   }
 
   @override
-  int get hashCode => Object.hash(name, bytes, height, width);
+  int get hashCode => Object.hash(
+        name,
+        bytes,
+        height,
+        width,
+        blurHash,
+      );
 
   @override
   bool operator ==(other) =>
@@ -47,5 +63,6 @@ class FFUploadedFile {
       name == other.name &&
       bytes == other.bytes &&
       height == other.height &&
-      width == other.width;
+      width == other.width &&
+      blurHash == other.blurHash;
 }
