@@ -1,91 +1,79 @@
 // Export pages
-export '/mobile/auth/onboarding/onboarding_widget.dart' show OnboardingWidget;
-export '/mobile/auth/sign_in/sign_in_widget.dart' show SignInWidget;
-export '/mobile/main/home_page/home_page_widget.dart' show HomePageWidget;
-export '/mobile/auth/sign_up/sign_up_widget.dart' show SignUpWidget;
-export '/mobile/auth/code/code_widget.dart' show CodeWidget;
-export '/mobile/auth/sign_up_code/sign_up_code_widget.dart'
-    show SignUpCodeWidget;
-export '/mobile/auth/sign_up_2/sign_up2_widget.dart' show SignUp2Widget;
-export '/mobile/booking/booking_page/booking_page_widget.dart'
-    show BookingPageWidget;
-export '/mobile/more/terms_of_use/terms_of_use_widget.dart'
-    show TermsOfUseWidget;
-export '/mobile/mynotes/my_notes/my_notes_widget.dart' show MyNotesWidget;
-export '/mobile/booking/current_booking_record/current_booking_record_widget.dart'
+export '/onboarding/onboarding_widget.dart' show OnboardingWidget;
+export '/sign_in/sign_in_widget.dart' show SignInWidget;
+export '/home_page/home_page_widget.dart' show HomePageWidget;
+export '/sign_up/sign_up_widget.dart' show SignUpWidget;
+export '/code/code_widget.dart' show CodeWidget;
+export '/sign_up_2/sign_up2_widget.dart' show SignUp2Widget;
+export '/sign_up_code/sign_up_code_widget.dart' show SignUpCodeWidget;
+export '/booking_page/booking_page_widget.dart' show BookingPageWidget;
+export '/terms_of_use/terms_of_use_widget.dart' show TermsOfUseWidget;
+export '/my_notes/my_notes_widget.dart' show MyNotesWidget;
+export '/current_booking_record/current_booking_record_widget.dart'
     show CurrentBookingRecordWidget;
-export '/mobile/more/about_company/about_company_widget.dart'
-    show AboutCompanyWidget;
-export '/mobile/profile/profile/profile_widget.dart' show ProfileWidget;
-export '/mobile/profile/select_city/select_city_widget.dart'
-    show SelectCityWidget;
-export '/mobile/profile/edit_profile/edit_profile_widget.dart'
-    show EditProfileWidget;
-export '/mobile/profile/favorites_car_washes/favorites_car_washes_widget.dart'
+export '/about_company/about_company_widget.dart' show AboutCompanyWidget;
+export '/profile/profile_widget.dart' show ProfileWidget;
+export '/select_city/select_city_widget.dart' show SelectCityWidget;
+export '/edit_profile/edit_profile_widget.dart' show EditProfileWidget;
+export '/favorites_car_washes/favorites_car_washes_widget.dart'
     show FavoritesCarWashesWidget;
-export '/w_e_b_admin/w_e_b_auth/admin_sign_in_email/admin_sign_in_email_widget.dart'
+export '/admin_sign_in_email/admin_sign_in_email_widget.dart'
     show AdminSignInEmailWidget;
-export '/w_e_b_admin/w_e_b_auth/admin_sign_up_email/admin_sign_up_email_widget.dart'
+export '/admin_sign_up_email/admin_sign_up_email_widget.dart'
     show AdminSignUpEmailWidget;
-export '/w_e_b_admin/w_e_b_auth/admin_forget_password/admin_forget_password_widget.dart'
+export '/admin_forget_password/admin_forget_password_widget.dart'
     show AdminForgetPasswordWidget;
-export '/w_e_b_admin/w_e_b_auth/admin_forget_password_2/admin_forget_password2_widget.dart'
+export '/admin_forget_password_2/admin_forget_password2_widget.dart'
     show AdminForgetPassword2Widget;
-export '/w_e_b_admin/w_e_b_main/admin_main/admin_main_widget.dart'
-    show AdminMainWidget;
-export '/w_e_b_admin/w_e_b_main/admin_reports/admin_reports_widget.dart'
-    show AdminReportsWidget;
-export '/w_e_b_admin/w_e_b_main/admin_office/admin_office_widget.dart'
-    show AdminOfficeWidget;
-export '/w_e_b_admin/w_e_b_main/admin_clients/admin_clients_widget.dart'
-    show AdminClientsWidget;
-export '/w_e_b_admin/w_e_b_main/admin_analytics/admin_analytics_widget.dart'
-    show AdminAnalyticsWidget;
-export '/w_e_b_admin/w_e_b_add_company/admin_add_company_1/admin_add_company1_widget.dart'
+export '/admin_main/admin_main_widget.dart' show AdminMainWidget;
+export '/admin_reports/admin_reports_widget.dart' show AdminReportsWidget;
+export '/admin_office/admin_office_widget.dart' show AdminOfficeWidget;
+export '/admin_clients/admin_clients_widget.dart' show AdminClientsWidget;
+export '/admin_analytics/admin_analytics_widget.dart' show AdminAnalyticsWidget;
+export '/admin_add_company_1/admin_add_company1_widget.dart'
     show AdminAddCompany1Widget;
-export '/w_e_b_admin/w_e_b_add_company/admin_add_company_3/admin_add_company3_widget.dart'
+export '/admin_add_company_3/admin_add_company3_widget.dart'
     show AdminAddCompany3Widget;
-export '/w_e_b_admin/w_e_b_add_company/admin_add_company_2/admin_add_company2_widget.dart'
+export '/admin_add_company_2/admin_add_company2_widget.dart'
     show AdminAddCompany2Widget;
-export '/w_e_b_admin/w_e_b_add_company/admin_add_company_4/admin_add_company4_widget.dart'
+export '/admin_add_company_4/admin_add_company4_widget.dart'
     show AdminAddCompany4Widget;
-export '/w_e_b_admin/w_e_b_add_company/admin_add_company_5/admin_add_company5_widget.dart'
+export '/admin_add_company_5/admin_add_company5_widget.dart'
     show AdminAddCompany5Widget;
-export '/w_e_b_admin/w_e_b_add_staff/admin_add_staff_2/admin_add_staff2_widget.dart'
+export '/admin_add_staff_2/admin_add_staff2_widget.dart'
     show AdminAddStaff2Widget;
-export '/w_e_b_admin/w_e_b_add_staff/admin_add_staff_3/admin_add_staff3_widget.dart'
+export '/admin_add_staff_3/admin_add_staff3_widget.dart'
     show AdminAddStaff3Widget;
-export '/w_e_b_admin/w_e_b_add_staff/admin_add_staff_1/admin_add_staff1_widget.dart'
+export '/admin_add_staff_1/admin_add_staff1_widget.dart'
     show AdminAddStaff1Widget;
-export '/w_e_b_super_admin/w_e_b_super_admin/super_admin_main/super_admin_main_widget.dart'
+export '/super_admin_main/super_admin_main_widget.dart'
     show SuperAdminMainWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin_moderation_company/super_admin_moderation/super_admin_moderation_widget.dart'
+export '/super_admin_moderation/super_admin_moderation_widget.dart'
     show SuperAdminModerationWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin_moderation_company/super_admin_moderation_1/super_admin_moderation1_widget.dart'
+export '/super_admin_moderation_1/super_admin_moderation1_widget.dart'
     show SuperAdminModeration1Widget;
-export '/w_e_b_super_admin/w_e_b_super_admin_moderation_company/super_admin_moderation_2/super_admin_moderation2_widget.dart'
+export '/super_admin_moderation_2/super_admin_moderation2_widget.dart'
     show SuperAdminModeration2Widget;
-export '/w_e_b_super_admin/w_e_b_super_admin_moderation_company/super_admin_moderation_3/super_admin_moderation3_widget.dart'
+export '/super_admin_moderation_3/super_admin_moderation3_widget.dart'
     show SuperAdminModeration3Widget;
-export '/w_e_b_super_admin/w_e_b_super_admin_moderation_company/super_admin_current_company/super_admin_current_company_widget.dart'
+export '/super_admin_current_company/super_admin_current_company_widget.dart'
     show SuperAdminCurrentCompanyWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin/super_admin_analytics/super_admin_analytics_widget.dart'
+export '/super_admin_analytics/super_admin_analytics_widget.dart'
     show SuperAdminAnalyticsWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin/super_admin_current_user/super_admin_current_user_widget.dart'
+export '/super_admin_current_user/super_admin_current_user_widget.dart'
     show SuperAdminCurrentUserWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin_moderation_promotions/super_admin_sales_moderation/super_admin_sales_moderation_widget.dart'
+export '/super_admin_sales_moderation/super_admin_sales_moderation_widget.dart'
     show SuperAdminSalesModerationWidget;
-export '/mobile/booking/booking_page_copy/booking_page_copy_widget.dart'
+export '/booking_page_copy/booking_page_copy_widget.dart'
     show BookingPageCopyWidget;
-export '/w_e_b_admin/w_e_b_main/admin_notifications/admin_notifications_widget.dart'
+export '/admin_notifications/admin_notifications_widget.dart'
     show AdminNotificationsWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin/super_admin_cancelled_bookings/super_admin_cancelled_bookings_widget.dart'
+export '/super_admin_cancelled_bookings/super_admin_cancelled_bookings_widget.dart'
     show SuperAdminCancelledBookingsWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin/super_admin_profile/super_admin_profile_widget.dart'
+export '/super_admin_profile/super_admin_profile_widget.dart'
     show SuperAdminProfileWidget;
-export '/w_e_b_admin/w_e_b_main/manager_profile/manager_profile_widget.dart'
-    show ManagerProfileWidget;
-export '/w_e_b_super_admin/w_e_b_super_admin/super_admin_users/super_admin_users_widget.dart'
+export '/manager_profile/manager_profile_widget.dart' show ManagerProfileWidget;
+export '/super_admin_users/super_admin_users_widget.dart'
     show SuperAdminUsersWidget;
 export '/admin_user_not_active/admin_user_not_active_widget.dart'
     show AdminUserNotActiveWidget;
